gjor kode mer lettleselig
